Fix __result type mismatch in Patch_Thing_Ingested
RimWorld 1.6 Thing.Ingested() returns float (System.Single), not int.
Changed __result from ref int to ref float to match the actual return type.

+++ Source/ShapeshifterFramework v1.6/ShapeshifterFramework/Patches/Patch_Thing_Ingested.cs
@@ -14,7 +14,7 @@ namespace ShapeshifterFramework.Patches
     internal static class Patch_Thing_Ingested
     {
         /// <summary>변신 중이거나 이데올로기 금기인 폰이 변신 약물을 섭취하면 차단.</summary>
-        static bool Prefix(Thing __instance, Pawn ingester, ref int __result)
+        static bool Prefix(Thing __instance, Pawn ingester, ref float __result)
         {
             if (ingester == null) return true;
             if (!ShapeshiftEligibility.HasShapeshiftOutcomeDoer(__instance.def)) return true;
@@ -26,7 +26,7 @@ namespace ShapeshifterFramework.Patches
             // 섭취 차단: 약물 소모 없이 메시지만 표시
             Messages.Message("SSF_Menu_Blocked".Translate(), ingester, MessageTypeDefOf.RejectInput, false);
             ingester.jobs?.EndCurrentJob(Verse.AI.JobCondition.Incompletable, false);
-            __result = 0;
+            __result = 0f;
             return false;
         }
     }
